refactor: FieldRenderer에서 prop에 대한 분기처리 작업 및 파일 분리
